refactor: move task-category-store imports to a new location
