+++ apps/web/src/services/ops-flow/components/TaskStatusTree.vue
@@ -11,7 +11,7 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 import TaskStatusList from '@/services/ops-flow/components/TaskStatusList.vue';
 import { TASK_STATUS_LABELS } from '@/services/ops-flow/constants/task-status-label-constant';
 import { useTaskCategoryPageStore } from '@/services/ops-flow/stores/admin/task-category-page-store';
-import { useTaskCategoryStore } from '@/services/ops-flow/stores/admin/task-category-store';
+import { useTaskCategoryStore } from '@/services/ops-flow/stores/task-category-store';
 
 const taskCategoryPageStore = useTaskCategoryPageStore();
 const taskCategoryStore = useTaskCategoryStore();
