+++ apps/web/src/services/ops-flow/stores/board-page-store.ts
@@ -5,7 +5,7 @@ import { defineStore } from 'pinia';
 
 import type { TaskCategoryModel } from '@/schema/opsflow/task-category/model';
 
-import { useTaskCategoryStore } from '@/services/ops-flow/stores/admin/task-category-store';
+import { useTaskCategoryStore } from '@/services/ops-flow/stores/task-category-store';
 
 interface UseBoardPageStoreGetters {
     currentCategory: ComputedRef<TaskCategoryModel|undefined>;
